simple setup/build shell script; calculate IP automatically for auth redirect
added a simple build script.
the ip is now calculated automagically so that rather than hardcoding one
it should use your outbound ip for the auth redirect stuff. this is only
a temporary work around for working on the project locally.

we need some kind of trigger for when the code-base is in production mode for
demos or something? im not sure how this would toy in a production server with
this little hack...