various bug fixes, closes #26 and closes #31
this is mostly duct tape kinda code and definitely needs cleaning and some mild re-works but the basis is there. ive implemented api routes for everything specified in 26, as well as implemented 31 so that all routes have a get/post/delete. by default it is assumed a route has a GET and nothing else unless specified.