recent activity changes; new activity_plan db table; rename db tables
when GLP's are unassigned they are now tracked in the recent activities
table thing.

when a GLP is assigned or un-assigned it is also stored/or deleted in a separate table
called 'active_plan'.

I've renamed some DB table names to be singular rather than plural as
per convention as well as added a new DB