cleaned up searches; searches now takes json query - closes #26;
cleaned up some of the api, there is now a parsing utility (which is currently not exhaustive...) for easily
doing an api request and parsing it.

the searches now take jquery post requests and use them properly, applying sorting options on glps, as well
as filtering for only showing glps, students, or for both to be searched.