classroom/student now uses the assignedglps hard request
this request means that all of the getglps are done on the server side rather than
the client side.

before we would perform a request, get some json back with all the glp id's, and then
we would request the getGlp(id) for each of the glps returned. this meant the client
was doing say 50 - 200 requests. this has now been put onto the server side squashing
it into a single request.