major style update, removed all non i18n phrases from code (hopefully)...
student analytics labels are changeed to black for the white background.
style update removing most of the orange borders.
removed all of the non i18n phrases from the code so it should just be awaits
moved some of the untranslated strings into the i18n config file.
changed white spacing to make things more balanced (margins are now equal, for example).

NOTE: this is not finished, there are a few things that still need styling